+++ libraries/ghci/GHCi/TH/Binary.hs
@@ -29,6 +29,7 @@ instance Binary TH.Stmt
 instance Binary TH.Pat
 instance Binary TH.Exp
 instance Binary TH.Dec
+instance Binary TH.Overlap
 instance Binary TH.Guard
 instance Binary TH.Body
 instance Binary TH.Match